Initial experiment with voice input
It seems that without ssl, each call to start() will request
permission for the microphone again. Luckily, github pages appears
to support https, so it should work nicely.

Input "completion" is somewhat slow, so we will probably want to
use `interimResults` for faster feedback: we can parse the interim
results and throw out any that don't match; if we find one that
matches, we go ahead and restart recognition to avoid dups