Begin support for parsing speech into commands!
It's a simple regex match. Given that ATC communication seems to
be fairly structured, this seems a reasonable approach. It works
pretty well so far, but definitely need the speed improvements
discussed in a previous commit. Might also be nice to have a way
to cancel a particular recognition (in case it started with something
unhelpful---though perhaps we should trim that off ourselves?)

Another nice thing would be to pause recognition while the radio
is speaking, since they wouldn't be able to hear you while they're
speaking on the radio anyway.