Begin working on a JSGF grammar for ATC
Our biggest problem here will be waypoints. Short of using the
nato-phonetic alphabet (which would certainly be easiest) I'm not
sure how to include them in a simple way that wouldn't be a pain
in the ass to add more (since you'd have to add the phonemes or
whatever for each one).

Adding the airline callsigns is not too bad since at worst it's
one phoneme per airline, but adding an airport probably involves
a whole mess of waypoints that will likely all need phonemes....